Update oci-build to use args+tags inputs as newline delimited
Github does weird things when injecting templated outputs into more complex YAML structures,
such as multiline strings. As a result, it did not like the 'split on semicolon, join on newline,
with JSON as intermediary' approach. This update moves from using semicolons to newlines for the
arg+tag delimiter, which is the lowest common denominator that won't get mangled by the GHA syntax