Remove pre-processing of build-args to avoid GHA bug
GHA YAML processing does not handle the 'multi-line string to variable back to multi-line string'
conversion properly. I cannot figure out what the problem is or why it's happening, but the combination
of providing a multiline string as a variable into a YAML field that is *not* specified as multiline
allows it to be parsed correctly.

However, there is no way to dynamically assemble a multiline string in a variable interpolation, so
there is no way to preprocess it and then inject it as a multiline value later. The result is that
we need to remove both the pre-processing and the multiline anchor (| in this case) so that the raw
input (specified as a multiline string) is interpolated into a non-multiline YAML value so that it
is parsed correctly by the engine.